GLOSSARY-53 Add ability to sync multilang Glossary with deepl glossary
* Rename class from translator to translation
